chore: moved getUserToken to kratos folder
