+++ core/api/src/app/authentication/login.ts
@@ -8,7 +8,6 @@ import {
 import { createAccountForDeviceAccount } from "@/app/accounts/create-account"
 
 import {
-  AuthenticationError,
   EmailUnverifiedError,
   IdentifierNotFoundError,
 } from "@/domain/authentication/errors"
@@ -58,7 +57,6 @@ import { ErrorLevel } from "@/domain/shared"
 import { consumeLimiter } from "@/services/rate-limit"
 import { RateLimitConfig } from "@/domain/rate-limit"
 import { RateLimiterExceededError } from "@/domain/rate-limit/errors"
-import { kratosAdmin } from "@/services/kratos/private"
 
 export const loginWithPhoneToken = async ({
   phone,
@@ -403,32 +401,3 @@ const checkDeviceLoginAttemptPerAppcheckJtiLimits = async (
     keyToConsume: appcheckJti,
   })
 
-export const getAuthTokenFromUserId = async (
-  userId: UserId,
-): Promise<AuthToken | AuthenticationError> => {
-  const { data } = await kratosAdmin.getIdentity({ id: userId })
-  let kratosResult:
-    | IAuthWithEmailPasswordlessService
-    | LoginWithPhoneNoPasswordSchemaResponse
-    | KratosError
-    | null = null
-
-  const phone = data?.traits?.phone
-  const email = data?.traits?.email
-
-  if (phone) {
-    const authService = AuthWithPhonePasswordlessService()
-    kratosResult = await authService.loginToken({ phone })
-  } else if (email) {
-    const emailAuthService = AuthWithEmailPasswordlessService()
-    kratosResult = await emailAuthService.loginToken({ email })
-  } else {
-    return new IdentifierNotFoundError()
-  }
-
-  if (kratosResult instanceof Error) {
-    return kratosResult
-  }
-
-  return kratosResult?.authToken
-}
