+++ core/api/src/services/kratos/identity.ts
@@ -4,6 +4,8 @@ import { KratosError, UnknownKratosError } from "./errors"
 import { kratosAdmin, toDomainIdentity } from "./private"
 
 import { IdentifierNotFoundError } from "@/domain/authentication/errors"
+import { AuthWithPhonePasswordlessService } from "./auth-phone-no-password"
+import { AuthWithEmailPasswordlessService } from "./auth-email-no-password"
 
 export const IdentityRepository = (): IIdentityRepository => {
   const getIdentity = async (
@@ -85,4 +87,34 @@ export const getNextPageToken = (link: string): string | undefined => {
   }
 
   return undefined
+}
+
+export const getAuthTokenFromUserId = async (
+  userId: UserId,
+): Promise<AuthToken | AuthenticationError> => {
+  const { data } = await kratosAdmin.getIdentity({ id: userId })
+  let kratosResult:
+    | IAuthWithEmailPasswordlessService
+    | LoginWithPhoneNoPasswordSchemaResponse
+    | KratosError
+    | null = null
+
+  const phone = data?.traits?.phone
+  const email = data?.traits?.email
+
+  if (phone) {
+    const authService = AuthWithPhonePasswordlessService()
+    kratosResult = await authService.loginToken({ phone })
+  } else if (email) {
+    const emailAuthService = AuthWithEmailPasswordlessService()
+    kratosResult = await emailAuthService.loginToken({ email })
+  } else {
+    return new IdentifierNotFoundError()
+  }
+
+  if (kratosResult instanceof Error) {
+    return kratosResult
+  }
+
+  return kratosResult?.authToken
 }
